gpt(eval): Add ResultMode strategies to parse/convert scores
The time finally came to separate this stuff. Also, I was forced to do
this in an OO way (instead of relying solely on enum/match) because the
binary part has no parameters, but the score does. If Python had
payloads in enums, this would've been fine, but it doesn't, so I need
the OO approach.