gpt(eval): Get result type from the prompt instead of the CLI
It doesn't make sense to get it from the CLI since each prompt already
"knows" its result type. I added it to the USER_PROMPTs dict, which
isn't particularly clean, but it should be good enough for now.

In the feature, I might move the prompts to text files. Maybe have the
result mode as a filename convention? Anyway, we're not there yet.