preprocess: Add script to process the CausalQA dataset
It outputs a format the graph algorithms can ingest.

PS: The `process_context` function likely needs more work. The context
from the input is weird. It's a single string containing multiple
articles and search results, so it needs to be split into its components
for our graph algorithms to work. This seems to work all right, but it
can't handle nested parentheses. I'll add some test for this specific
function later on and try to solve this.