graph(baseline): Keep log of all interactions per item ID
This means I can inspect the output later to see everything that we sent
to the API and what it replied back, which can be useful for debugging
and verification.

However, this means that I now have to keep some state in the Client to
be able to keep track of this. I tried another version where the calls
returned a log alongside the result, but it looked ugly because I had to
explicitly merge the logs later.

A problem with this approach is that now everything needs to know about
the item_id so I can pass it to the API call so it can update is
internal log state. It was either that or manually getting the log after
an iteration with a given item and explicity resetting the log state,
which was too much for me.

This approach doesn't spark joy, so I might still come back to an
explicit log handling. This will do for now.