graph(baseline): Implement my own Graph instead of using networkx
I wasn't using nx for anything, really, just for basic graph
representation, and it turned out that it wasn't great for what I wanted
to do. This simple Graph class does everything I wanted.

I'm _slightly_ disappointed with the abstraction, though: I can't make
the objects immutable. The attributes themselves are, but nothing stops
anyone from re-assigning them. I tried using dataclass for this, but it
was too much boilerplate. The alternative here is making them private
with only public properties to get them, but that still felt like too
much boilerplate for something so simple. I'll leave this for now. It
probably won't ever be an issue anyway.