graph(baseline): Add steps to merge nodes and relations before combining
The implementation is rather simple: give a list to the LLM, ask it to
merge and create a mapping, them create new graph by applying the
mapping.

Right now, there's a lot in common between the node and relation merging
functions, but I'm not going to do much to reduce that since it looks
coincidental rather than inherent.