graph(simple): Add another baseline that does the simplest thing possible
It just gets all of the texts, appends them without any transformation
whatsoever to the prompt, and asks the model for the answer. All in one
go.

Unfortunately, it's currently better than the graph version while being
much faster and cheaper. It sets the bare minimum I need to achieve with
the new approaches.