metrics: Re-write the metrics based on CausalQA's evaluation script
It was doing anything _too_ different, but I felt it would be the best
to replicate exactly what they were doing just in case there were any
differences.

I needed to do some refactoring to make it work, but nothing that should
alter the final value. Maybe for the sentence-level since they were
doing their own accumulation of the metrics per sample alongside the
overall dataset and I went with just using a single-sample dataset
approach for that, but I didn't check the per sample results and they
don't matter that much anyway.