graph(baseline): Remove restriction on number of graphs
Some datasets don't have contexts at all, so I don't mind them failing
(e.g. GooAQ), but some have contexts for _most_ of the entries except
for some, and I think this should work with that.

This change should be enough to make the algorithm work with cases with
no contexts (which would yield no graphs here) by just giving the
remaining parts of it empty graphs to handle.