graph: Refactor common code into separate module
`simple.py` was initially a copy of `baseline.py`, so there was a lot of
redundant code there. This is the first refactor to decrease that.

No, I don't have a better name than `common`.

Also, there will be more. The `main` function still has a lot of exact
copied code, which I'm thinking of how to deduplicate that. I won't be
today, though. I'm already glad I did this first pass today.